move clickhouse to protocoltest dir similar to other db test dir
