+++ lib/srv/db/clickhouse/protocoltest/test.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package clickhouse
+package protocoltest
 
 import (
 	"context"
